1-strncat.c

+++ 0x06-pointers_arrays_strings/1-strncat.c
@@ -10,7 +10,7 @@
 
 char *_strncat(char *dest, char *src, int n)
 {
-	int m = n;
+	size_t m = n;
 
 	strncat(dest, src, m)
 	return (dest);
